requirements.txt
anyio==3.7.0
certifi==2023.5.7
dnspython==2.3.0
et-xmlfile==1.1.0
flake8==6.0.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
mccabe==0.7.0
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.2
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
six==1.16.0
sniffio==1.3.0
telegram-text==0.1.2
tzdata==2023.3
